Finally, all components pass all unit tests, as demonstrated by running test(all). Additional tests need to be written. Also, a change is coming to the trajectory interface that may break a few things. In particular, trajectories will soon be required to predict into the future if evaluated beyond the upper bound of their time domain.

+++ trunk/components/+Default/LinearKalman.m
@@ -78,7 +78,7 @@ classdef LinearKalman < tom.Optimizer & Default.DefaultConfig
     end
     
     function num=numSolutions(this)
-      num=numel(this.dynamicModel);
+      num=uint32(numel(this.dynamicModel));
     end
     
     function xEst=getSolution(this,k)
